refactor(rcml): move vertical email templates to their vendor packages
Moves the three single-consumer vertical email template modules out of
@rule-io/rcml into the vendor packages that actually use them, shrinking
rcml/src/ (non-email/) from ~5760 LoC to ~3639 LoC (-37%). External
@rule-io/sdk consumers are unaffected because sdk also re-exports the
vendor packages.

File moves (git-renamed so diff survives)
-----------------------------------------
- packages/rcml/src/hospitality-templates.ts
    → packages/vendor-bookzen/src/hospitality-templates.ts
- packages/rcml/src/ecommerce-templates.ts
    → packages/vendor-shopify/src/ecommerce-templates.ts
- packages/rcml/src/generic-templates.ts
    → packages/vendor-shopify/src/generic-templates.ts

In each moved file the `./types.js` / `./brand-template.js` / `./elements.js`
/ `./utils.js` imports were rewritten to `@rule-io/rcml` so the vendor
packages reach their shared dependencies the same way every other external
consumer does.

Public API surface
------------------
- packages/rcml/src/index.ts drops the three vertical-template export
  blocks (5 + 4 + 1 functions + 5 + 4 + 1 option types); leaves a comment
  pointing at the new homes.
- packages/rcml/src/index.ts promotes three previously-internal
  brand-template helpers (createLogoSection, createGreetingSection,
  createCtaSection) to the public API so the moved files can reach them
  across the package boundary.
- packages/vendor-bookzen/src/index.ts re-exports the five hospitality
  templates + their option types.
- packages/vendor-shopify/src/index.ts re-exports the four e-commerce
  templates + createWelcomeEmail + their option types.

Consumer wiring
---------------
- packages/vendor-bookzen/src/automations.ts imports the hospitality
  builders from ./hospitality-templates.js instead of @rule-io/rcml.
- packages/vendor-shopify/src/automations.ts does the same for
  ./ecommerce-templates.js and ./generic-templates.js.

Test-file split
---------------
packages/rcml/tests/templates.test.ts shrank from 3208 → 1149 lines and
now contains only brand-template / validateCustomFields / barrel-exports /
withTemplateContext suites. The vertical-template describes moved with
their source files:
  - packages/vendor-bookzen/tests/hospitality-templates.test.ts (new, 34 tests)
  - packages/vendor-shopify/tests/ecommerce-templates.test.ts   (new, 83 tests)
These new files re-use the per-vendor helpers.ts fixtures (TEST_BRAND_STYLE,
assertValidRCMLDocument, docToString) that were already in place.

Lint clean-up (ride-along)
--------------------------
Removed six now-unused eslint-disable directives:
  - packages/rcml/src/email/content/converter/convert.test.ts (3 sites — trimmed
    no-unsafe-argument sub-directive from three combined comments)
  - packages/rcml/src/email/content/json-validator/validate.test.ts (same)
  - packages/rcml/src/email/content/json-validator/json-schema.ts
    (dropped explicit-module-boundary-types disable above buildJsonSchema)
  - packages/rcml/src/email/validate-rcml-json.ts (dropped
    no-duplicate-type-constituents block; the re-export pattern no longer
    trips that rule).

Verification
------------
npx nx run-many -t build  --projects=rcml,vendor-bookzen,vendor-shopify,vendor-samfora,sdk ✓
npx nx run-many -t eslint:lint --projects=rcml,vendor-bookzen,vendor-shopify ✓ (0 errors)
npx nx run-many -t vitest:test  --projects=rcml,vendor-bookzen,vendor-shopify,vendor-samfora
  → rcml 955 + vendor-bookzen 34 + vendor-shopify 83 + vendor-samfora 33
  = 1105 tests passing, 0 failing.
